refactoring code and alignment
