+++ src/main/java/com/springjwt/common/aspect/LoggerAspect.java
@@ -46,9 +46,4 @@ public class LoggerAspect {
         log.info("After executing Service method. Returned: {}", returnValue);
     }
 
-
-
-
-
-
 }
